Separação em classes StatementPrinter
