+++ TheatricalPlayersRefactoringKata.Tests/StatementPrinterTests.cs
@@ -27,7 +27,7 @@ public class StatementPrinterTests
         );
 
         StatementPrinter statementPrinter = new StatementPrinter();
-        var result = statementPrinter.Print(invoice, plays);
+        var result = statementPrinter.Print(invoice, plays, "txt");
 
         Approvals.Verify(result);
     }
@@ -58,7 +58,7 @@ public class StatementPrinterTests
         );
 
         StatementPrinter statementPrinter = new StatementPrinter();
-        var result = statementPrinter.Print(invoice, plays);
+        var result = statementPrinter.Print(invoice, plays, "txt");
 
         Approvals.Verify(result);
     }
@@ -89,7 +89,7 @@ public class StatementPrinterTests
         );
 
         StatementPrinter statementPrinter = new StatementPrinter();
-        var result = statementPrinter.PrintXml(invoice, plays);
+        var result = statementPrinter.Print(invoice, plays, "xml");
 
         Approvals.Verify(result);
     }
