+++ TheatricalPlayersRefactoringKata/IStatementPrinter.cs
@@ -1,0 +1,9 @@
+﻿using System.Collections.Generic;
+
+namespace TheatricalPlayersRefactoringKata
+{
+    public interface IStatementPrinter
+    {
+        string Print(Invoice invoice, Dictionary<string, Play> plays);
+    }
+}
